webhooks: Remove the legacy GitHub integration.
github-services is no longer available in production. This also
allows us to finally remove authenticated_api_view.
