+++ zerver/tests/test_decorators.py
@@ -30,7 +30,6 @@ from zerver.lib.request import \
     RequestVariableConversionError, RequestConfusingParmsError
 from zerver.decorator import (
     api_key_only_webhook_view,
-    authenticated_api_view,
     authenticated_json_view,
     authenticated_rest_api_view,
     authenticated_uploads_api_view,
@@ -460,51 +459,6 @@ class SkipRateLimitingTest(ZulipTestCase):
             self.assertTrue(rate_limit_mock.called)
 
 class DecoratorLoggingTestCase(ZulipTestCase):
-    def test_authenticated_api_view_logging(self) -> None:
-        @authenticated_api_view(is_webhook=True)
-        def my_webhook_raises_exception(request: HttpRequest, user_profile: UserProfile) -> None:
-            raise Exception("raised by webhook function")
-
-        webhook_bot_email = 'webhook-bot@zulip.com'
-        webhook_bot_realm = get_realm('zulip')
-        webhook_bot = get_user(webhook_bot_email, webhook_bot_realm)
-        webhook_bot_api_key = get_api_key(webhook_bot)
-
-        request = HostRequestMock()
-        request.method = 'POST'
-        request.POST['api_key'] = webhook_bot_api_key
-        request.POST['email'] = webhook_bot_email
-        request.host = "zulip.testserver"
-
-        with mock.patch('zerver.decorator.webhook_logger.exception') as mock_exception:
-            with self.assertRaisesRegex(Exception, "raised by webhook function"):
-                request.body = '{}'
-                request.POST['payload'] = '{}'
-                request.content_type = 'text/plain'
-                my_webhook_raises_exception(request)  # type: ignore # mypy doesn't seem to apply the decorator
-
-            message = """
-user: {email} ({realm})
-client: {client_name}
-URL: {path_info}
-content_type: {content_type}
-custom_http_headers:
-{custom_headers}
-body:
-
-{body}
-                """
-            message = message.strip(' ')
-            mock_exception.assert_called_with(message.format(
-                email=webhook_bot_email,
-                realm=webhook_bot_realm.string_id,
-                client_name='Unspecified',
-                path_info=request.META.get('PATH_INFO'),
-                content_type=request.content_type,
-                custom_headers=None,
-                body=request.body,
-            ))
-
     def test_authenticated_rest_api_view_logging(self) -> None:
         @authenticated_rest_api_view(webhook_client_name="ClientName")
         def my_webhook_raises_exception(request: HttpRequest, user_profile: UserProfile) -> None:
