--- zerver/webhooks/github_legacy/fixtures/v2_push_multiple_committers.json
@@ -1,185 +1,0 @@
-{
-  "event": "push",
-  "payload": {
-    "forced": false,
-    "compare": "https://github.com/zbenjamin/zulip-test/compare/4f9adc4777d5...b95449196980",
-    "after": "b95449196980507f08209bdfdc4f1d611689b7a8",
-    "repository": {
-      "fork": false,
-      "watchers": 0,
-      "description": "Test repository",
-      "has_downloads": true,
-      "url": "https://github.com/zbenjamin/zulip-test",
-      "stargazers": 0,
-      "created_at": 1354060018,
-      "private": false,
-      "name": "zulip-test",
-      "pushed_at": 1377639818,
-      "owner": {
-        "name": "zbenjamin",
-        "email": "zev-github@strangersgate.com"
-      },
-      "has_wiki": true,
-      "open_issues": 0,
-      "has_issues": true,
-      "forks": 0,
-      "master_branch": "master",
-      "id": 6893087,
-      "size": 66
-    },
-    "created": false,
-    "deleted": false,
-    "commits": [
-      {
-        "committer": {
-          "username": "zbenjamin",
-          "name": "Zev Benjamin",
-          "email": "zev@mit.edu"
-        },
-        "added": [
-          "baz"
-        ],
-        "author": {
-          "username": "zbenjamin",
-          "name": "Zev Benjamin",
-          "email": "zev@mit.edu"
-        },
-        "distinct": true,
-        "timestamp": "2013-08-27T14:42:23-07:00",
-        "modified": [],
-        "url": "https://github.com/zbenjamin/zulip-test/commit/48c329a0b68a9a379ff195ee3f1c1f4ab0b2a89e",
-        "message": "Add baz\n\nWe need it for testing",
-        "removed": [],
-        "id": "48c329a0b68a9a379ff195ee3f1c1f4ab0b2a89e"
-      },
-      {
-        "committer": {
-          "username": "tomasz",
-          "name": "Tomasz Kolek",
-          "email": "zev@mit.edu"
-        },
-        "added": [],
-        "author": {
-          "username": "tomasz",
-          "name": "Tomasz Kolek",
-          "email": "zev@mit.edu"
-        },
-        "distinct": true,
-        "timestamp": "2013-08-27T14:42:23-07:00",
-        "modified": [],
-        "url": "https://github.com/zbenjamin/zulip-test/commit/48c329a0b68a9a379ff195ee3f1c1f4ab0b2a89e",
-        "message": "Add baz\n\nWe need it for testing",
-        "removed": [],
-        "id": "48c329a0b68a9a379ff195ee3f1c1f4ab0b2a89e"
-      },
-        {
-        "committer": {
-          "username": "tomasz",
-          "name": "Tomasz Kolek",
-          "email": "zev@mit.edu"
-        },
-        "added": [],
-        "author": {
-          "username": "tomasz",
-          "name": "Tomasz Kolek",
-          "email": "zev@mit.edu"
-        },
-        "distinct": true,
-        "timestamp": "2013-08-27T14:42:23-07:00",
-        "modified": [],
-        "url": "https://github.com/zbenjamin/zulip-test/commit/48c329a0b68a9a379ff195ee3f1c1f4ab0b2a89e",
-        "message": "Add baz\n\nWe need it for testing",
-        "removed": [],
-        "id": "48c329a0b68a9a379ff195ee3f1c1f4ab0b2a89e"
-      },
-        {
-        "committer": {
-          "username": "tomasz",
-          "name": "Tomasz Kolek",
-          "email": "zev@mit.edu"
-        },
-        "added": [],
-        "author": {
-          "username": "tomasz",
-          "name": "Tomasz Kolek",
-          "email": "zev@mit.edu"
-        },
-        "distinct": true,
-        "timestamp": "2013-08-27T14:42:23-07:00",
-        "modified": [],
-        "url": "https://github.com/zbenjamin/zulip-test/commit/48c329a0b68a9a379ff195ee3f1c1f4ab0b2a89e",
-        "message": "Add baz\n\nWe need it for testing",
-        "removed": [],
-        "id": "48c329a0b68a9a379ff195ee3f1c1f4ab0b2a89e"
-      },
-      {
-        "committer": {
-          "username": "baxthehacker",
-          "name": "Ben Simon",
-          "email": "zev@mit.edu"
-        },
-        "added": [],
-        "author": {
-          "username": "baxthehacker",
-          "name": "Ben Simon",
-          "email": "zev@mit.edu"
-        },
-        "distinct": true,
-        "timestamp": "2013-08-27T14:42:23-07:00",
-        "modified": [],
-        "url": "https://github.com/zbenjamin/zulip-test/commit/48c329a0b68a9a379ff195ee3f1c1f4ab0b2a89e",
-        "message": "Add baz\n\nWe need it for testing",
-        "removed": [],
-        "id": "48c329a0b68a9a379ff195ee3f1c1f4ab0b2a89e"
-      },
-        {
-        "committer": {
-          "username": "baxthehacker",
-          "name": "Ben Simon",
-          "email": "zev@mit.edu"
-        },
-        "added": [],
-        "author": {
-          "username": "baxthehacker",
-          "name": "Ben Simon",
-          "email": "zev@mit.edu"
-        },
-        "distinct": true,
-        "timestamp": "2013-08-27T14:42:23-07:00",
-        "modified": [],
-        "url": "https://github.com/zbenjamin/zulip-test/commit/48c329a0b68a9a379ff195ee3f1c1f4ab0b2a89e",
-        "message": "Add baz\n\nWe need it for testing",
-        "removed": [],
-        "id": "48c329a0b68a9a379ff195ee3f1c1f4ab0b2a89e"
-      }
-    ],
-    "pusher": {
-      "name": "zbenjamin",
-      "email": "zev-github@strangersgate.com"
-    },
-    "head_commit": {
-      "committer": {
-        "username": "zbenjamin",
-        "name": "Zev Benjamin",
-        "email": "zev@mit.edu"
-      },
-      "added": [],
-      "author": {
-        "username": "zbenjamin",
-        "name": "Zev Benjamin",
-        "email": "zev@mit.edu"
-      },
-      "distinct": true,
-      "timestamp": "2013-08-27T14:43:09-07:00",
-      "modified": [
-        "baz"
-      ],
-      "url": "https://github.com/zbenjamin/zulip-test/commit/b95449196980507f08209bdfdc4f1d611689b7a8",
-      "message": "Final edit to baz, I swear\n\nI mean, really, could you leave it at \"1, 2\"?",
-      "removed": [],
-      "id": "b95449196980507f08209bdfdc4f1d611689b7a8"
-    },
-    "ref": "refs/heads/master",
-    "before": "4f9adc4777d54ccb7ac105c98fe393c078a33f53"
-  }
-}
